requirements.txt
streamlit
pdfminer.six
spacy
scikit-learn
